# docs/management-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CodeSearch Management API",
    "version": "0.1.0",
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "description": "Local management server (default bind). Use 0.0.0.0 when started with --public.",
      "variables": {
        "port": {
          "default": "8676",
          "description": "Management port passed to `codesearch serve --mgmt-port` (default 8676)."
        }
      }
    }
  ],
...
  "paths": {
...
    "/api/context/{symbol}": {
      "parameters": [
        {
          "name": "symbol",
          "in": "path",
          "required": true,
          "description": "Symbol name, or regex when `regex=true`.",
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "repository",
          "in": "query",
          "required": false,
          "description": "Restrict to a repository (name or UUID).",
          "schema": {
            "type": "string"
          }
        },
        {
          "name": "regex",
          "in": "query",
          "required": false,
          "description": "Interpret `{symbol}` as a regular expression.",
          "schema": {
            "type": "boolean",
            "default": false
          }
        }
      ],
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Callers and callees of a symbol",
        "description": "Returns the serialized `SymbolContext` domain value (callers up to entry points, callees down to leaves).",
        "operationId": "context",
        "responses": {
          "200": {
            "description": "Symbol context.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymbolContext"
                }
              }
            }
          },
          "404": {
            "description": "The symbol was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/features": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Entry-point execution features by criticality",
        "operationId": "features",
        "parameters": [
          {
            "name": "repository",
            "in": "query",
            "required": false,
            "description": "Repository to analyse (name or UUID). Omit to auto-detect.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Maximum number of features to return.",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Execution features sorted by criticality.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "repository_id",
                    "count",
                    "features"
                  ],
                  "properties": {
                    "repository_id": {
                      "type": "string"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ExecutionFeature"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/couplings": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Coupling elements holding fragile communities together",
        "description": "Files/symbols or dependencies whose removal would split a Leiden community into two latent sub-blocks (the hub-like dependency / modularity-violation smell). Returns the serialized `CouplingReport` domain value.",
        "operationId": "couplings",
        "parameters": [
          {
            "name": "repository",
            "in": "query",
            "required": false,
            "description": "Repository to analyse (name or UUID). Omit to auto-detect.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "description": "Which graph to analyse: the file-dependency graph or the symbol call graph.",
            "schema": {
              "type": "string",
              "enum": [
                "file",
                "symbol"
              ],
              "default": "file"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The coupling report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CouplingReport"
                }
              }
            }
          },
          "400": {
            "description": "Invalid level parameter.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels": {
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Cross-service channel links",
        "description": "Matched producer\u2192consumer links (Kafka topics, HTTP routes, MQTT/AMQP topics, gRPC methods). Returns the serialized `ChannelLinkReport` domain value. With no `repository`, scoped to the current namespace's repositories.",
        "operationId": "channels",
        "parameters": [
          {
            "name": "repository",
            "in": "query",
            "required": false,
            "description": "Restrict to specific repositories (name or UUID). Repeatable.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "description": "Filter by protocol.",
            "schema": {
              "type": "string",
              "enum": [
                "kafka",
                "http",
                "mqtt",
                "amqp",
                "grpc"
              ]
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "description": "Drop edges below this confidence (0.0\u20131.0).",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "exclude_channel",
            "in": "query",
            "required": false,
            "description": "Exclude channels matching these globs. Repeatable.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "include_tests",
            "in": "query",
            "required": false,
            "description": "Include endpoints from test files (excluded by default).",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The channel-link report.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelLinkReport"
                }
              }
            }
          },
          "400": {
            "description": "Unknown protocol.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "A named repository was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Uniform error body returned with a 4xx/5xx status.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable failure message."
          }
        }
      },
...
      "SymbolContext": {
        "type": "object",
        "description": "Serialized `SymbolContext` domain value: callers (entry points \u2192 symbol) and callees (symbol \u2192 leaves) as trees.",
        "additionalProperties": true
      },
...
      "ExecutionFeature": {
        "type": "object",
        "description": "Serialized execution-feature domain value: an entry point with its criticality score and contributing factors.",
        "additionalProperties": true
      },
...
      "CouplingReport": {
        "type": "object",
        "description": "Serialized `CouplingReport` domain value: per internally-fragile Leiden community, its two latent sub-blocks and the ablation-verified coupling elements (nodes/edges) holding them together, with split probabilities and the resolution range each controls.",
        "additionalProperties": true
      },
      "ChannelLinkReport": {
        "type": "object",
        "description": "Serialized `ChannelLinkReport` domain value: matched producer\u2192consumer channel links with protocol, channel name, endpoints, and confidence.",
        "additionalProperties": true
      },
...
    }
  }
}